Some renaming and added open/close modal for team logic
