+++ src/html/footer.html
@@ -9,7 +9,13 @@
         <svg class="footer-section__icon" width="14" height="13">
           <use href="/src/images/decorative/sprite.svg#icon-Vector-heart"></use>
         </svg>
-        by &nbsp;<a href="/" class="footer-section__acсent"> GoIT Students</a>
+        by &nbsp;<a
+          href="/"
+          class="footer-section__acсent"
+          id="open-team-modal"
+        >
+          GoIT Students</a
+        >
       </span>
     </div>
   </div>
